user changes

+++ src/main/java/ghebrat/bes/theresFoodAtHome/user/UserRepo.java
@@ -2,10 +2,9 @@ package ghebrat.bes.theresFoodAtHome.user;
 
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface UserRepo extends CrudRepository<User, Long> {
+public interface UserRepo extends CrudRepository<User, Integer> {
 
     Optional<User> findByEmail(String email);
